+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 - Added /rfuchangelog command to view the mod changelog of your current version
 - Added custom fishing keybind overrides for hotbar slots 1-9, left click, and right click
 - Added tracking and HUD displays for power orb fluxes (Mana Flux, Overflux, Plasmaflux) with dynamic rarity coloring and stat boosts
+- Added :angry:, :fire:, and :scream: emojis
 
 ### Changes
-- Made the titanoboa and scuttler emojis bigger
 - Bobber display, personal cap display, and blizzard timer now automatically hide when the fishing session is paused
 
 ### Fixes